Se organizo la consulta de productos, ademas se agrego query para las variaciones de un producto
