+++ graphql/schemas/querys/index.js
@@ -1,9 +1,11 @@
-let WooCommerceProduct = require('./WooCommerce/wooCommerceProduct.query');
+const WooCommerceProduct = require('./WooCommerce/wooCommerceProduct.query');
+const WooCommerceProductVariation = require('./WooCommerce/wooCommerceProductVariation.query');
 const ShopifyProducts = require('./Shopify/shopifyProduct.query.js');
 
 module.exports = {
     public: {
         WooCommerceProduct,
-        ShopifyProducts
+        WooCommerceProductVariation,
+        ShopifyProducts,
     }
 }
