+++ graphql/schemas/types/wooCommerce/Product/WooCommerceTaxType.js
@@ -1,9 +1,7 @@
 const {
     GraphQLObjectType,
     GraphQLString,
-    GraphQLBoolean,
     GraphQLInt,
-    GraphQLList
 } = require('graphql');
 
 let WooCommerceTaxType = new GraphQLObjectType({
