+++ graphql/schemas/types/wooCommerce/Product/wooCommerceCategoryType.js
