+++ graphql/schemas/types/wooCommerce/Product/wooCommerceProduct.type.js
@@ -7,12 +7,8 @@ const {
 } = require('graphql');
 
 const WooCommerceCategoryType = require('./wooCommerceCategoryType');
-const WooCommerceImageType = require('./WooCommerceImageType');
-const WooCommerceProductVariationType = require('./WooCommerceProductVariationType');
 const WoocommerceTaxType =  require('./WooCommerceTaxType');
 const stripHtml = require("string-strip-html");
-
-const { getVariations, getCategories, getTax } = require('../../../../controllers/WooCommerce.controller');
 
 let WooCommerceProductType = new GraphQLObjectType({
     name: 'WooCommerceProductType',
@@ -22,9 +18,6 @@ let WooCommerceProductType = new GraphQLObjectType({
         return obj.id.toString();
       }},
       reference:{ type:GraphQLString, resolve:(obj, args, context, info)=>{
-        if(obj.sku === 'CM021401'){
-          console.log(obj)
-        }
         return obj.sku
       }}, //Referencia del Producto
       description:{ type:GraphQLString, resolve:(obj, args, context, info)=>{
@@ -57,7 +50,7 @@ let WooCommerceProductType = new GraphQLObjectType({
             return null;
         }
         
-      }}, 
+      }},
       mainCategory: { type:WooCommerceCategoryType, resolve:(obj, args, context, info)=>{
         return obj.categories.length > 0 ? obj.categories[0] : null
       }}, //Categoria Principal del Producto
@@ -89,53 +82,6 @@ let WooCommerceProductType = new GraphQLObjectType({
       } }, //Peso del Empaque del Producto
       categories: { type:new GraphQLList(WooCommerceCategoryType), resolve:(obj, args, context, info)=>{
         return obj.categories || [];
-      }},
-      images:{ type:new GraphQLList(WooCommerceImageType), resolve:(obj, args, context, info)=>{
-        return obj.images
-      }},
-      variations:{ type:new GraphQLList(WooCommerceProductVariationType), resolve:(obj, args, context, info)=>{
-          let getGender = (p)=>{
-            if(( p.attributes &&  p.attributes.length > 0 )){
-              let attrs = obj.attributes;
-              let gender = attrs.filter(o=>(o.name.toLowerCase() === 'gender' || o.name.toLowerCase() === 'genero' || o.name === 'Género'));
-    
-              if(gender.length > 0)
-                return gender[0].options[0];
-              else
-                return null;
-            }
-            return null;
-          }
-
-          let getSize = (p)=>{
-            if(( p.attributes &&  p.attributes.length > 0 )){
-              let attrs = p.attributes;
-              let size = attrs.filter(o=>(o.name.toLowerCase() === 'talla' || o.name.toLowerCase() === 'tamaño' || o.name.toLowerCase() === 'size'));
-    
-              if(size.length > 0)
-                return size[0].options[0];
-              else
-                return 'único';
-            }
-    
-            return 'único';
-          }
-
-          if(!obj.variations || obj.variations.length === 0){
-            let defaultVariation = {
-                sku:obj.sku,
-                ean13:obj.ean13 || '',
-                upc:obj.upc || '', 
-                price:obj.price ? parseInt(obj.price == "" ? 0 : obj.price) : 0,
-                gender:getGender(obj), //Género para el cual aplica el producto (Masculino, Femenino, Unisex, Niños, Niñas)
-                talla:getSize(obj), //Género para el cual aplica el producto (Masculino, Femenino, Unisex, Niños, Niñas)
-                stock_quantity:obj.stock_quantity || 0
-            }
-            
-            return [defaultVariation];
-          }
-
-        return getVariations(context.req, obj.id);
       }},
     }),
 });
