+++ graphql/schemas/types/wooCommerce/ProductVariations/wooCommerceProductVariation.type.js
@@ -1,0 +1,50 @@
+const {
+    GraphQLObjectType,
+    GraphQLString,
+    GraphQLInt,
+} = require('graphql');
+
+let WooCommerceProductVariationType = new GraphQLObjectType({
+  name: 'WooCommerceProductVariationType',
+  fields: () => ({
+    supplierreference:{ type:GraphQLString },
+    ean13:{type:GraphQLInt},
+    upc:{type:GraphQLInt}, 
+    price:{ type:GraphQLInt, resolve:(obj, args, context, info)=>{
+      return obj.price ? parseInt(obj.price == "" ? 0 : obj.price) : 0
+    }},
+    gender:{ type:GraphQLString, resolve:(obj, args, context, info)=>{
+      if(( obj.attributes &&  obj.attributes.length > 0 )){
+        let attrs = obj.attributes;
+        let gender = attrs.filter(o=>(o.name.toLowerCase() === 'gender' || o.name.toLowerCase() === 'genero' || o.name.toLowerCase() === 'género'));
+        if(gender.length > 0)
+          return gender[0].option;
+        else
+          return null;
+      }else{
+        return obj.gender;
+      }
+    }}, //Género para el cual aplica el producto (Masculino, Femenino, Unisex, Niños, Niñas)
+    talla:{ type:GraphQLString, resolve:(obj, args, context, info)=>{
+      if(( obj.attributes &&  obj.attributes.length > 0 )){
+        let attrs = obj.attributes;
+        let size = attrs.filter(o=>(o.name.toLowerCase() === 'talla' || o.name.toLowerCase() === 'tamaño' || o.name.toLowerCase() === 'size'));
+
+        if(size.length > 0)
+          return size[0].option;
+        else
+          return null;
+      }else{
+        return obj.talla;
+      }
+    }}, //Género para el cual aplica el producto (Masculino, Femenino, Unisex, Niños, Niñas)
+    quantity:{ type:GraphQLInt, resolve:(obj, args, context, info)=>{
+      return obj.stock_quantity || 0
+    }},
+    reference:{ type:GraphQLString, resolve:(obj, args, context, info)=>{
+      return obj.sku
+    }}
+  }),
+});
+
+module.exports = WooCommerceProductVariationType;
